Fix type casting to avoid resetting job time limit implicitly to (uint16_t) NO_VAL

+++ src/api/init_msg.c
@@ -96,7 +96,7 @@ void slurm_init_job_desc_msg(job_desc_msg_t * job_desc_msg)
 	job_desc_msg->argc        = 0;
 	job_desc_msg->shared      = (uint16_t) NO_VAL;
 	job_desc_msg->task_dist   = (uint16_t) NO_VAL;
-	job_desc_msg->time_limit  = (uint16_t) NO_VAL;
+	job_desc_msg->time_limit  = NO_VAL;
 	job_desc_msg->num_procs   = NO_VAL;
 	job_desc_msg->max_nodes   = NO_VAL;
 	job_desc_msg->min_nodes   = NO_VAL;
